fix unlimit calc endPoint (qint32)(qint64)

+++ src/wlgmachine.cpp
@@ -3487,7 +3487,7 @@ case WLElementTraj::line:  {
 
                               foreach(WLGDrive *mD,getGDrives()){
                               if(!mD->getAxis()) continue;
-                              ePos+=round(ME.data.line.endPoint.get(mD->getName())/mD->dimension());
+                              ePos+=(qint32)((qint64)round(ME.data.line.endPoint.get(mD->getName())/mD->dimension()));
                               indexs+=i;
                               i++;
                               }
@@ -3514,8 +3514,8 @@ case WLElementTraj::uline:   {
 
                              foreach(WLGDrive *mD,getGDrives()){
                               if(!mD->getAxis()) continue;
-                              ePos+=round(ME.data.uline.endPoint.get(mD->getName())/mD->dimension());
-                              mPos+=round(ME.data.uline.midPoint.get(mD->getName())/mD->dimension());
+                              ePos+=(qint32)((qint64)round(ME.data.uline.endPoint.get(mD->getName())/mD->dimension()));
+                              mPos+=(qint32)((qint64)round(ME.data.uline.midPoint.get(mD->getName())/mD->dimension()));
                               indexs+=i;
                               i++;
                               }
@@ -3548,8 +3548,8 @@ case WLElementTraj::arc: {
 
                           foreach(WLGDrive *mD,getGDrives()){
                            if(!mD->getAxis()) continue;
-                           ePos+=round(ME.data.arc.endPoint.get(mD->getName())/mD->dimension());
-                           cPos+=round(ME.data.arc.centerPoint.get(mD->getName())/mD->dimension());
+                           ePos+=(qint32)((qint64)round(ME.data.arc.endPoint.get(mD->getName())/mD->dimension()));
+                           cPos+=(qint32)((qint64)round(ME.data.arc.centerPoint.get(mD->getName())/mD->dimension()));
                            indexs+=i;
                            i++;
                            }
